[1-splash-screen] Refactoring name of application
