+++ .gitignore
@@ -12,3 +12,4 @@
 /captures
 .externalNativeBuild
 .cxx
+/.idea/
