+++ app/build.gradle
@@ -7,7 +7,7 @@ android {
     buildToolsVersion "29.0.2"
 
     defaultConfig {
-        applicationId "com.olmi.android.memms"
+        applicationId "com.olmi.android.memes"
         minSdkVersion 23
         targetSdkVersion 29
         versionCode 1
