+++ app/src/androidTest/java/com/olmi/android/memes/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.olmi.android.memms
+package com.olmi.android.memes
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.olmi.android.memms", appContext.packageName)
+        assertEquals("com.olmi.android.memes", appContext.packageName)
     }
 }
