+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Memms</string>
+    <string name="app_name">Memems</string>
 </resources>
