+++ settings.gradle
@@ -1,2 +1,2 @@
-rootProject.name='Memms'
+rootProject.name='Memes'
 include ':app'
